Creating Ui for stating pipeline
