+++ src/webview/pipeline/app/const.ts
@@ -1,0 +1,12 @@
+/*-----------------------------------------------------------------------------------------------
+ *  Copyright (c) Red Hat, Inc. All rights reserved.
+ *  Licensed under the MIT License. See LICENSE file in the project root for license information.
+ *-----------------------------------------------------------------------------------------------*/
+
+
+export enum VolumeTypes {
+  EmptyDirectory = 'Empty Directory',
+  ConfigMap = 'Config Map',
+  Secret = 'Secret',
+  PVC = 'PVC',
+}
